[rev] added error check for args

+++ script/timefree.sh
@@ -22,19 +22,26 @@ programInfo=$1
 duration=$2
 fromTime=$3
 
+[ "$(echo ${fromTime} | grep -Eo [0-9]{14})" = "" ] && fromTime="";
+
+# in case usage (2),(3) (programInfo=radiko URL), set stationID and fromTime
 if [ -n "$(echo ${programInfo} | grep -Eo [0-9]{14})" ];then
 
 	# URL copied from browser
 	stationID=`echo ${programInfo} | grep -o sid.*\& | sed "y/\&/=/" |  cut -d= -f2`
 
 	# URL copied from radiko function
-	[ -n "${stationID}" ] && stationID=`echo ${programInfo} | grep -o "\#\!/ts.*/" | cut -d/ -f3` 
+	[ "${stationID}" = "" ] && stationID=`echo ${programInfo} | grep -o "\#\!/ts.*/" | cut -d/ -f3`
 
+	# if fromTime is specified by both programInfo and fromTime, override by programInfo
 	fromTime=`echo ${programInfo} | grep -Eo [0-9]{14}`
 	
 else
 	stationID=${programInfo}
 fi
+
+# if fromTime is blank, exit
+[ "${fromTime}" = "" ] && { echo -e "fromTime is blank. check radiko URL or startTime args."; exit; }
 
 toTime=`echo ${fromTime} |
        	sed -e "s/[0-9][0-9]/ &/g" -e "s/ \([0-9][0-9]\) /\1/" | # separate startTime into YYYY_MM_DD_HH_MM_SS 
@@ -50,6 +57,9 @@ else
 	mail=$(cat "${ROOTPATH}/mail")
 fi
 
+
+# if duration is less than 1, exit
+[ "${duration}" -lt 1 ] && { echo -e "duration must be more than 1."; exit; }
 
 # set temporary workdir
 workdir=/tmp/radiko-$$
@@ -85,31 +95,6 @@ bluetoothaddr=$(hcitool con | grep -Eo "[0-9A-F:]{9,}")
 
 ffmpegOption=" -y -protocol_whitelist file,pipe,crypto -f concat -i - -c copy -loglevel fatal "
 
-#auth1_fms="${workdir}/free_${channel}.auth1_fms"
-#auth2_fms="${workdir}/free_${channel}.auth2_fms"
-
-#loginfile="${workdir}/login"
-#cookiefile=${workdir}/cookie.txt
-#playerfile=${workdir}/player.swf
-#keyfile=${workdir}/authkey.png
-
-#if [ -z "$3" ]; then
-#	firstAACfile="/tmp/${firstName}.aac"
-#	secondAACfile="/tmp/${secondName}.aac"
-#	restAACfile="/tmp/${restName}.aac"
-#fi
-
-
-#baseinput="${workdir}/baseinput.m3u8"
-#grepinput="${workdir}/grepinput.m3u8"
-
-#wholeAAC="${workdir}/wholeAAC.txt"
-
-#------------------------------------------------------------
-#------------------------------------------------------------
-#------------------------------------------------------------
-#------------------------------------------------------------
-
 # check enpass / set password
 checkEnpass $radikoCredentialPath
 
@@ -126,10 +111,13 @@ getAuth2 $workdir
 getStreamURL $workdir $stationID $fromTime $toTime
 
 
+[ "$(grep -c ^ ${baseinput})" -lt 2 ] && { echo -e "baseinput not found. check stationID, dateTime is correct."; exit; }
+
 stream_url=$(grep radiko ${baseinput})
 rm ${baseinput}
 
 wget -q $stream_url -O - | grep "radiko.jp/sound" > ${wholeAACList}
+
 
 # download first 12 AAC files
 cat ${wholeAACList} | head -n 12 | xargs -P 0 -L 1 wget -q -nc -P ${tmpDirPath}
